- refactor DocumentList moving some code to utils
